Handle failing allocation in createPrerenderCanvas
Allocating the bitmap pixels for the PrerenderedInval
object might fail due to insufficient memory, bitmap
size is negative or size doesn't fit in 32 bits.
A failing allocation will later result in a crash
when drawing operations are performed on the bitmap.

The crash is avoided by checking the return value
from SkBitmap::allocPixels and return 0 if it fails.

Change-Id: I2215dd6cb403b28e2bd5a0aa6cb6d5c8bdc6c173