Pause drawing when not visible
Even when a page was in the background and paused contentDraw() was
still called, e.g. caused by animated gifs or javascript updates.
This could cause congestion on the WebViewCoreThread when background
pages were updated frequently.

Now contentDraw() will not be called as long as WebViewCore is
paused, and contentDraw() will be called on resume. This also has
the the side-effect of suspending the GIF animation loop.

Change-Id: I8a5df65e7cbc136438e616951f667f1f073ef504